More physics changes and 1 bug found
Summary:

• Fixed jump animations to work with other changes.
• Found a bug with turning around, detailed below.

One bug I’m noticing—and this was true before I did anything as well—if
you hit left and then immediately hit right, like within a frame of
each other, you start to fall through the platform, but snap back on
top of it an instant later. I didn’t run into this with my old
character controller, so I’m not sure if that has to do with
FixedUpdate or what. It seems like it’s potentially problematic and
also it’s noticeable as the character jitters. Just figured I’d mention
it, though I’ll be looking into it as well.