Collision check fix
I didn’t like how the .OverlapCircle worked for collision so I replaced
it with .OverlapArea. I also gave the player a box collider with no
circle collider to make it feel more like old games where you could
stand with only a pixel on a platform and not fall off.

One thing I’m struggling with now, though, is the guy seems to not be
grounded on part of the hill and the area after it. Not sure what’s up
with that.