Bug update
Ok just to make a note about the bug in question here—that you stop
being grounded on the hill to the right and the flat part beyond—I
noticed that the point where this happens is at player position y=0.5.
I’ve been banging my head against this all day and can’t figure out
what the hell is going on so I’m just giving up for now.