Restructuring hit interaction stuff.
A whole interaction controller thing was a little bit overkill. Should
be able to do similar stuff with standard SendMessage stuff.

Adding crouchAttack to pick up the balast. Now you can attack "a" to cut
the balast loose, crouch attack "down+a" to pick it up, and secondary
fire "s" to use it (currently just logs "ok, using it" or something to
indicate us throwing it or whatever).